[#51] refactor: 카테고리 캐시를 위한 카테고리 맵 서비스 생성, 구조 리팩토링
카테고리 맵 서비스 <- 카테고리 서비스 <- 카테고리 컨트롤러
