+++ kyumall-client/src/main/java/com/kyumall/kyumallclient/product/CategoryService.java
@@ -1,35 +1,51 @@
 package com.kyumall.kyumallclient.product;
 
-import com.kyumall.kyumallcommon.exception.ErrorCode;
-import com.kyumall.kyumallcommon.exception.KyumallException;
+import com.kyumall.kyumallclient.product.dto.CategoryDto;
 import com.kyumall.kyumallcommon.product.entity.Category;
 import com.kyumall.kyumallcommon.product.repository.CategoryRepository;
-import com.kyumall.kyumallcommon.product.vo.CategoryStatus;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
 public class CategoryService {
+  private final CategoryMapService categoryMapService;
   private final CategoryRepository categoryRepository;
 
-  //캐시는 히트율이 중요함 id 별로 카테고리를 캐시하지 말고, 전체 카테고리를 캐시해 둘 것
-  public Category findCategoryById(Long categoryId) {
-    return categoryRepository.findById(categoryId)
-        .orElseThrow(() -> new KyumallException(ErrorCode.CATEGORY_NOT_EXISTS));
+  public List<CategoryDto> getAllCategories() {
+    Map<Long, List<Category>> categoryGroupingByParent = categoryMapService.findCategoryGroupingByParent();
+    return convertToCategoryHierarchy(categoryGroupingByParent);
+  }
+
+
+  /**
+   * 카테고리 Map을 계층형 구조의 List로 변경
+   * @param groupingByParentId
+   * @return
+   */
+  private List<CategoryDto> convertToCategoryHierarchy(Map<Long, List<Category>> groupingByParentId) {
+    List<CategoryDto> rootCategories = groupingByParentId.get(0L).stream().map(CategoryDto::from)
+        .toList();
+    addSubCategories(rootCategories, groupingByParentId);
+    return rootCategories;
   }
 
   /**
-   * 전체 카테고리를 조회하여 parentId 로 group by 한 Map 을 만듭니다.
-   * @return all category grouping by parent id
+   * 재귀 호출로 카테고리의 서브 카테고리를 계층형으로 추가합니다.
+   * @param categoryDtos
+   * @param groupingByParentId
    */
-  @Cacheable(value = "categoryMap", key = "#root.methodName")
-  public Map<Long, List<Category>> findCategoryGroupingByParent() {
-    List<Category> allCategory = categoryRepository.findAllByStatus(CategoryStatus.INUSE);
-    return allCategory.stream().collect(Collectors.groupingBy(Category::getParentId));
+  private void addSubCategories(List<CategoryDto> categoryDtos, Map<Long, List<Category>> groupingByParentId) {
+    categoryDtos.stream().forEach(
+        categoryDto -> {
+          List<CategoryDto> subCategories = groupingByParentId.getOrDefault(categoryDto.getId(), new ArrayList<>())
+              .stream().map(CategoryDto::from).toList();
+          categoryDto.setSubCategories(subCategories);
+          addSubCategories(subCategories, groupingByParentId);
+        }
+    );
   }
 }
